refactor: rename datepickerPropTypes
